refactor(tabs): optimize CSS
