+++ src/components/tabs/themes/shared/tab/tab.bootstrap.scss
@@ -5,27 +5,17 @@ $theme: $bootstrap;
 
 :host {
     transition: none;
-    background: var-get($theme, 'item-background');
-    color: var-get($theme, 'item-text-color');
-    border-radius: var-get($theme, 'border-radius');
     border-end-end-radius: 0;
     border-end-start-radius: 0;
 }
 
 :host(:hover) {
-    background: var-get($theme, 'item-hover-background');
-    color: var-get($theme, 'item-hover-color');
     box-shadow: inset 0 0 0 rem(1px) var-get($theme, 'border-color--hover');
 }
 
 :host([selected]) {
-    color: var-get($theme, 'item-active-color');
     position: relative;
     box-shadow: inset 0 0 0 rem(1px) var-get($theme, 'border-color');
-
-    ::slotted(igc-icon) {
-        color: var-get($theme, 'item-active-icon-color');
-    }
 
     &::after {
         content: '';
@@ -57,12 +47,4 @@ $theme: $bootstrap;
     &::after {
         display: none;
     }
-}
-
-:host([selected]:focus-within) {
-    color: var-get($theme, 'item-active-color');
-}
-
-:host([disabled]) {
-    color: var-get($theme, 'item-disabled-color');
 }
