+++ src/components/tabs/themes/shared/tab/tab.common.scss
