+++ src/components/tabs/themes/shared/tab/tab.fluent.scss
@@ -4,37 +4,12 @@
 $theme: $fluent;
 
 :host {
-    background: var-get($theme, 'item-background');
-    color: var-get($theme, 'item-text-color');
-    border-radius: var-get($theme, 'border-radius');
-
     ::slotted(igc-icon) {
         color: var-get($theme, 'item-icon-color');
     }
 }
 
-:host(:hover),
-:host(:focus-within) {
-    background: var-get($theme, 'item-hover-background');
-    color: var-get($theme, 'item-hover-color');
-}
-
-:host([selected]) {
-    color: var-get($theme, 'item-active-color');
-
-    ::slotted(igc-icon) {
-        color: var-get($theme, 'item-active-icon-color');
-    }
-}
-
-:host([selected]:hover),
-:host([selected]:focus-within) {
-    color: var-get($theme, 'item-active-color');
-}
-
 :host([disabled]) {
-    color: var-get($theme, 'item-disabled-color');
-
     ::slotted(igc-icon) {
         color: var-get($theme, 'item-disabled-icon-color');
     }
