--- src/components/tabs/themes/shared/tab/tab.indigo.scss
@@ -1,33 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $indigo;
-
-:host {
-    background: var-get($theme, 'item-background');
-    color: var-get($theme, 'item-text-color');
-    border-radius: var-get($theme, 'border-radius');
-}
-
-:host(:hover),
-:host(:focus-within) {
-    background: var-get($theme, 'item-hover-background');
-    color: var-get($theme, 'item-hover-color');
-}
-
-:host([selected]) {
-    color: var-get($theme, 'item-active-color');
-
-    ::slotted(igc-icon) {
-        color: var-get($theme, 'item-active-icon-color');
-    }
-}
-
-:host([selected]:hover),
-:host([selected]:focus-within) {
-    color: var-get($theme, 'item-active-color');
-}
-
-:host([disabled]) {
-    color: var-get($theme, 'item-disabled-color');
-}
