+++ src/components/tabs/themes/shared/tabs/tabs.bootstrap.scss
@@ -3,10 +3,6 @@
 
 $theme: $bootstrap;
 
-[part='headers'] {
-    background: var-get($theme, 'item-background');
-}
-
 [part='headers-wrapper'] {
     &::after {
         content: '';
@@ -18,29 +14,6 @@ $theme: $bootstrap;
         background: var-get($theme, 'border-color');
         z-index: 0;
     }
-}
-
-[part='start-scroll-button'],
-[part='end-scroll-button'] {
-    background: var-get($theme, 'button-background');
-
-    &:hover,
-    &:focus {
-        background: var-get($theme, 'button-hover-background');
-    }
-}
-
-igc-icon-button::part(base) {
-    color: var-get($theme, 'button-color');
-
-    &:hover,
-    &:focus-within {
-        color: var-get($theme, 'button-hover-color');
-    }
-}
-
-igc-icon-button[disabled]::part(base) {
-    color: var-get($theme, 'button-disabled-color');
 }
 
 [part='selected-indicator'] {
