+++ src/components/tabs/themes/shared/tabs/tabs.common.scss
