+++ src/components/tabs/themes/shared/tabs/tabs.fluent.scss
@@ -3,38 +3,12 @@
 
 $theme: $fluent;
 
-[part='headers'] {
-    background: var-get($theme, 'item-background');
-}
-
-[part='selected-indicator'] {
-    background: var-get($theme, 'indicator-color');
-}
-
-[part='start-scroll-button'],
-[part='end-scroll-button'] {
-    background: var-get($theme, 'button-background');
-
-    &:hover,
-    &:focus {
-        background: var-get($theme, 'button-hover-background');
-    }
-}
-
 igc-icon-button::part(base) {
-    color: var-get($theme, 'button-color');
-
     igc-icon {
         color: var-get($theme, 'button-color');
-    }
-
-    &:hover,
-    &:focus-within {
-        color: var-get($theme, 'button-hover-color');
     }
 }
 
 igc-icon-button[disabled]::part(base) {
-    color: var-get($theme, 'button-disabled-color');
     background: inherit;
 }
