+++ src/components/tabs/themes/tab-themes.ts
@@ -2,38 +2,37 @@ import { css } from 'lit';
 
 // Shared Styles
 import { styles as bootstrap } from './shared/tab/tab.bootstrap.css.js';
+import { styles as shared } from './shared/tab/tab.common.css.js';
 import { styles as fluent } from './shared/tab/tab.fluent.css.js';
-import { styles as indigo } from './shared/tab/tab.indigo.css.js';
-import { styles as material } from './shared/tab/tab.material.css.js';
 import { Themes } from '../../../theming/types.js';
 
 const light = {
   bootstrap: css`
-    ${bootstrap}
+    ${shared} ${bootstrap}
   `,
   material: css`
-    ${material}
+    ${shared}
   `,
   fluent: css`
-    ${fluent}
+    ${shared} ${fluent}
   `,
   indigo: css`
-    ${indigo}
+    ${shared}
   `,
 };
 
 const dark = {
   bootstrap: css`
-    ${bootstrap}
+    ${shared} ${bootstrap}
   `,
   material: css`
-    ${material}
+    ${shared}
   `,
   fluent: css`
-    ${fluent}
+    ${shared} ${fluent}
   `,
   indigo: css`
-    ${indigo}
+    ${shared}
   `,
 };
