+++ src/components/tabs/themes/tabs-themes.ts
@@ -12,38 +12,37 @@ import { styles as indigoLight } from './light/tab.indigo.css.js';
 import { styles as materialLight } from './light/tab.material.css.js';
 // Shared Styles
 import { styles as bootstrap } from './shared/tabs/tabs.bootstrap.css.js';
+import { styles as shared } from './shared/tabs/tabs.common.css.js';
 import { styles as fluent } from './shared/tabs/tabs.fluent.css.js';
-import { styles as indigo } from './shared/tabs/tabs.indigo.css.js';
-import { styles as material } from './shared/tabs/tabs.material.css.js';
 import { Themes } from '../../../theming/types.js';
 
 const light = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapLight}
+    ${shared} ${bootstrap} ${bootstrapLight}
   `,
   material: css`
-    ${material} ${materialLight}
+    ${shared} ${materialLight}
   `,
   fluent: css`
-    ${fluent} ${fluentLight}
+    ${shared} ${fluent} ${fluentLight}
   `,
   indigo: css`
-    ${indigo} ${indigoLight}
+    ${shared} ${indigoLight}
   `,
 };
 
 const dark = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapDark}
+    ${shared} ${bootstrap} ${bootstrapDark}
   `,
   material: css`
-    ${material} ${materialDark}
+    ${shared} ${materialDark}
   `,
   fluent: css`
-    ${fluent} ${fluentDark}
+    ${shared} ${fluent} ${fluentDark}
   `,
   indigo: css`
-    ${indigo} ${indigoDark}
+    ${shared} ${indigoDark}
   `,
 };
